corrected composer.json
