+++ composer.json
@@ -1,5 +1,5 @@
 {
-  "name": "dreadkopp/HTML_OutputOptimizer",
+  "name": "dreadkopp/html_outputoptimizer",
   "description": "beautify, minify and optimize your HTML and images",
   "type": "library",
   "license": "MIT",
@@ -10,7 +10,7 @@
     }
   ],
   "require": {
-    "php": ">=5.6",
+    "php": "^7|^8",
     "ext-imagick" : "*",
     "ext-curl" : "*",
     "predis/predis": "^1.1",
